Move param and buffer logic to separate modules
